refactor: extract embedded attempt runner helpers
