+++ src/agents/embedded-agent-runner/run/attempt.run-decisions.test.ts
@@ -1,0 +1,92 @@
+import { describe, expect, it } from "vitest";
+import {
+  resolveAttemptStreamAuthProfileId,
+  resolveAttemptToolPolicyMessageProvider,
+  resolveEmbeddedAttemptSessionWriteLockOptions,
+  resolveUnknownToolGuardThreshold,
+  shouldRunLlmOutputHooksForAttempt,
+} from "./attempt.run-decisions.js";
+
+describe("resolveEmbeddedAttemptSessionWriteLockOptions", () => {
+  it("bounds post-prompt session lock max hold to compaction timeout instead of run timeout", () => {
+    const options = resolveEmbeddedAttemptSessionWriteLockOptions({
+      config: {},
+      compactionTimeoutMs: 600_000,
+      env: {},
+    });
+
+    expect(options.maxHoldMs).toBe(720_000);
+  });
+});
+
+describe("resolveAttemptStreamAuthProfileId", () => {
+  it("uses only the runtime-forwarded auth profile for stream provenance", () => {
+    expect(
+      resolveAttemptStreamAuthProfileId({
+        authProfileId: "openai-codex:raw-session-profile",
+        runtimePlan: {
+          auth: {
+            forwardedAuthProfileId: "openai-codex:forwarded-profile",
+          },
+        } as never,
+      }),
+    ).toBe("openai-codex:forwarded-profile");
+
+    expect(
+      resolveAttemptStreamAuthProfileId({
+        authProfileId: "openai:non-forwarded-profile",
+        runtimePlan: {
+          auth: {},
+        } as never,
+      }),
+    ).toBeUndefined();
+  });
+});
+
+describe("resolveAttemptToolPolicyMessageProvider", () => {
+  it("prefers explicit tool-policy provider over transport channel", () => {
+    expect(
+      resolveAttemptToolPolicyMessageProvider({
+        messageChannel: "discord",
+        messageProvider: "discord-voice",
+      }),
+    ).toBe("discord-voice");
+  });
+
+  it("falls back to message channel when provider is omitted", () => {
+    expect(resolveAttemptToolPolicyMessageProvider({ messageChannel: "discord" })).toBe("discord");
+  });
+});
+
+describe("shouldRunLlmOutputHooksForAttempt", () => {
+  it("skips llm_output after before_agent_run blocks before model submission", () => {
+    expect(shouldRunLlmOutputHooksForAttempt({ promptErrorSource: "hook:before_agent_run" })).toBe(
+      false,
+    );
+    expect(shouldRunLlmOutputHooksForAttempt({ promptErrorSource: "prompt" })).toBe(true);
+    expect(shouldRunLlmOutputHooksForAttempt({ promptErrorSource: null })).toBe(true);
+  });
+});
+
+describe("resolveUnknownToolGuardThreshold", () => {
+  it("returns the default threshold when no loop-detection config is provided", () => {
+    expect(resolveUnknownToolGuardThreshold(undefined)).toBe(10);
+    expect(resolveUnknownToolGuardThreshold({})).toBe(10);
+  });
+
+  it("stays on even when tools.loopDetection.enabled is false", () => {
+    expect(resolveUnknownToolGuardThreshold({ enabled: false })).toBe(10);
+    expect(resolveUnknownToolGuardThreshold({ enabled: false, unknownToolThreshold: 3 })).toBe(3);
+  });
+
+  it("uses positive integer thresholds and floors fractions", () => {
+    expect(resolveUnknownToolGuardThreshold({ enabled: true, unknownToolThreshold: 4 })).toBe(4);
+    expect(resolveUnknownToolGuardThreshold({ unknownToolThreshold: 3.7 })).toBe(3);
+  });
+
+  it("falls back to the default threshold when the override is non-positive", () => {
+    expect(resolveUnknownToolGuardThreshold({ unknownToolThreshold: 0 })).toBe(10);
+    expect(resolveUnknownToolGuardThreshold({ unknownToolThreshold: -5 })).toBe(10);
+    expect(resolveUnknownToolGuardThreshold({ unknownToolThreshold: Number.NaN })).toBe(10);
+  });
+});
